Move null view sanitization into VectorUtils
Fold ViewVectorSanitizer into VectorUtils.sanitizeNullViewSlots alongside the
other VectorSchemaRoot helpers; the Stitcher calls it before feeding its output.
